requirements.txt
streamlit==1.28.1
pymongo==4.6.0
pandas==2.1.3
plotly==5.17.0
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
openai>=1.3.0
anthropic>=0.8.0
tenacity>=8.2.0